Add initial project with unity 2022.3.22f1
This should be enough to run unity 2022.3.22f1, while I could add others at
this time I will not use them so cba.

A lot of this could be moved into functions to support multiple versions, but
since there is only one version there is no point.

I have been hacking on this outside of git for a while, so this represents a
lot of work I had not added to source control.